[utils] QmlInstantEngine: fix spurious consecutive reloads issue
Fix the problem with the application taking up all the CPU after a few reloads while using a Qt3D Scene.
When a file changes, wait for a few milliseconds before adding it back to the watch  list. This prevents the system to reload the engine twice for a same fileChanged event (happens with some text editors).