Clear clipboard's content when exiting
Workaround for a bug in QClipboard that occurs when the clipboard
has been used within the app and its content exceeds a certain size
on X11/XCB.
This issue will hold up the app when exiting and is present in Qt5
versions.

With this workaround, the content of the clipboard will be lost when
exiting, but the app will exit normally.