[qt6] Viewer2D: perform the automated auto-fit when the width is updated
The signals sent when an image is updated used to be firstly widthChanged()
then heightChanged(). It seems this order has been reversed with Qt6.4.
This requires updating the slot on which the auto-fit is performed because
both the height and width need to be known for the auto-fit to be correct.