[qt6] Viewer3D: remove "PointSize" RenderState
Using PointSize as a RenderState triggers a "RHI Unhandled render
state class Qt3DRender::Render::PointSize" warning and does not work.
It is removed from the 3D viewer and instead, the "fixedPointSize"
property is directly sent to QmlAlembic: when it is set to true, the
point size becomes literally fixed (at size 0.005f), and the behaviour
slightly differs from the one we had in Qt5, where this property set to
true led to the PointSize RenderState using the SizeMode::Fixed status;
this translated into the fixed point size being editable through the
slider, with a noticable difference in size compared to when
"fixedPointSize" was set to false.