[Viewer] Viewer2D: Load and unload SfmStats components explicitly
The SfmStatsView and SfmGlobalStats components used to be loaded once
and for all with "Component.onCompleted". With the upgrade to Qt 5.14,
it is needed to load and explicitly unload the component (by resetting its
source) for the statistics to always be correctly displayed. Otherwise,
they can be loaded and viewed only once per session; after being shown for
the first time, they are set to null and never reloaded.

This commit uses the same type of loading/unloading as all the other
components in the 2D Viewer.