Set GLTF as the default output format in Texturing
Note that GLTF generates an additional .bin file.
For this, I had to include .* in the output file pattern.
An alternative was to add an additional output parameter to be
enabled only when GLTF is selected (like the MTL file).
Also note that with this .* pattern, the MTL output is redundant.