[ui] Open a warning dialog when attempting to submit an unsaved project
A project needs to be saved in order to be submitted to the renderfarm.
Prior to this commit, attempting to submit an unsaved project would not
prompt any dialog in the UI despite the exception that was thrown on the
Python side.

A warning dialog is now opened to let the user know that the project
cannot be submitted until it is saved.