[core] Correctly parse status in version names when it exists
Version objects were meant to be used only with numerical characters,
separated by periods. Now that the version of Meshroom can also contain
its status ("develop" when it is in develop mode), the Version class
needs to be able to parse the status string as well as the version
components when it exists.

The status is handled as a separate string, which is returned in addition
to the tuple of integers when the whole version string is parsed in the
first place.

This prevents issues when opening files that have been saved with a version
like "x.x.x-develop".