[nodes] StructureFromMotion: Add new inputs parameters
Add two new input parameters, `nbFirstUnstableCameras` which sets the
number of cameras for which a bundle adjustment is performed every time
a camera is added, and `maxImagesPerGroup` which sets the maximum number
of cameras that can be added before the bundle adjustment has to be
performed again.

By default, the bundle adjustment will be performed every time one of
the first 30 cameras is added, as computations are not too expensive
while there are few cameras. Past that number, the bundle adjustment will
only be performed when up to 30 cameras have been added.

The node's version is updated accordingly.