[core] Handle `UidConflict` compatibility issues in node factory
The `UidConflict` is not detected per se within the node factory as a
node's UID cannot be computed accurately from there, since its expressions
(which may impact the UID computation) have not been resolved yet.

Instead, the `UidConflict` is detected externally, and the information
is transferred to the node factory, which will create a corresponding
compatibility node with the regular process.

The created `UidConflict` compatibility node will be handled by the
Compatibility Manager like any other compatibility node.