[nodes] HDR Fusion: Correctly detect number of brackets
The detection of the number of brackets used to only work in a case where
there was a single dataset / a single camera intrinsics. If two datasets
with the same number of brackets were provided, the detection was failing
because we expected the exposure levels to be uniform across all the
images.

If more than one dataset is provided, there is no guarantee that the
exposure groups will be identical although the number of brackets is the
same.

The inputs are however sorted, and the shutter speeds are expected to be
decreasing, meaning that a shutter speed N superior to a shutter speed N-1
indicates a new group. In the same manner, ISO or aperture values that
change from one input to the next one indicate a new group.

For the comparison between exposure levels to be valid, the aperture,
shutter speed and ISO values need to be stored in tuples as floats instead
of strings.