[nodes] LdrToHdrSampling: Exclude outliers from size computation
Outliers are now supported for the HDR fusion nodes and are excluded
from the computations as soon as they have been detected. However,
the chunks' size computation still includes them, as it simply uses the
number of detected brackets and the total number of input images.

With this commit, the detected outliers are excluded from the chunks' size
computation, thus preventing any error that might occur because there
are too many chunks compared to the number of images that actually need
to be processed.