[nodes] PanoramaPostProcessing: Add attributes to change the outputs' names
Add input String parameters to determine the name of the generated outputs.
These attributes include:
- the name of the generated output panorama
- the name of the generated output panorama preview
The base folder for these outputs remains the node's cache, only the
name of the outputs themselves can be updated.

The output attribute about the downscaled levels of the panorama is also
modified with the new name of the pyramid levels, which is a concatenation
of the generated panorama's name and the size of the current level.